# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.PointSaleCreate:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  schemas.PointSaleUpdate:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
...
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is a API for Club Norte
  termsOfService: http://swagger.io/terms/
  title: APP Club Norte
  version: "1.0"
paths:
...
  /v1/point_sale/create:
    post:
      consumes:
      - application/json
      description: PointSaleCreate
      parameters:
      - description: PointSaleCreate
        in: body
        name: point_sale_create
        required: true
        schema:
          $ref: '#/definitions/schemas.PointSaleCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: PointSaleCreate
      tags:
      - PointSale
  /v1/point_sale/delete/{id}:
    delete:
      consumes:
      - application/json
      description: PointSaleDelete
      parameters:
      - description: ID Point Sale
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: PointSaleDelete
      tags:
      - PointSale
  /v1/point_sale/get/{id}:
    get:
      consumes:
      - application/json
      description: PointSaleGet
      parameters:
      - description: ID Point Sale
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: PointSaleGet
      tags:
      - PointSale
  /v1/point_sale/get_all:
    get:
      consumes:
      - application/json
      description: PointSaleGetAll
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: PointSaleGetAll
      tags:
      - PointSale
  /v1/point_sale/update:
    put:
      consumes:
      - application/json
      description: PointSaleUpdate
      parameters:
      - description: PointSaleUpdate
        in: body
        name: point_sale_update
        required: true
        schema:
          $ref: '#/definitions/schemas.PointSaleUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: PointSaleUpdate
      tags:
      - PointSale
...
securityDefinitions:
  CookieAuth:
    description: 'Enter your JWT token here. Example: "eyJhbGciOiJIUz..."'
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"